consumer_api/requirements.txt
selenium
beautifulsoup4
kafka-python
influxdb-client